improved conversions
