+++ modules/config.py
@@ -206,7 +206,7 @@ class Config:
             try:                                self.tmdb["apikey"] = check_for_attribute(self.data, "apikey", parent="tmdb", throw=True)
             except Failed as e:                 raise Failed(e)
             self.tmdb["language"] = check_for_attribute(self.data, "language", parent="tmdb", default="en")
-            self.TMDb = TMDbAPI(self.tmdb)
+            self.TMDb = TMDbAPI(self, self.tmdb)
             logger.info(f"TMDb Connection {'Failed' if self.TMDb is None else 'Successful'}")
         else:
             raise Failed("Config Error: tmdb attribute not found")
@@ -755,194 +755,21 @@ class Config:
         for i, item in enumerate(items, 1):
             length = util.print_return(length, f"Processing: {i}/{len(items)} {item.title}")
             try:
-                id_type, main_id = self.get_id(item, library, length)
+                id_type, main_id = self.Arms.get_id(item, library, length)
             except BadRequest:
                 util.print_stacktrace()
                 util.print_end(length, f"{'Cache | ! |' if self.Cache else 'Mapping Error:'} | {item.guid} for {item.title} not found")
                 continue
-            if isinstance(main_id, list):
-                if id_type == "movie":
-                    for m in main_id:
-                        if m in movie_map:
-                            movie_map[m].append(item.ratingKey)
-                        else:
-                            movie_map[m] = [item.ratingKey]
-                elif id_type == "show":
-                    for m in main_id:
-                        if m in show_map:
-                            show_map[m].append(item.ratingKey)
-                        else:
-                            show_map[m] = [item.ratingKey]
-            else:
-                if id_type == "movie":
-                    if main_id in movie_map:
-                        movie_map[main_id].append(item.ratingKey)
-                    else:
-                        movie_map[main_id] = [item.ratingKey]
-                elif id_type == "show":
-                    if main_id in show_map:
-                        show_map[main_id].append(item.ratingKey)
-                    else:
-                        show_map[main_id] = [item.ratingKey]
+            if not isinstance(main_id, list):
+                main_id = [main_id]
+            if id_type == "movie":
+                for m in main_id:
+                    if m in movie_map:              movie_map[m].append(item.ratingKey)
+                    else:                           movie_map[m] = [item.ratingKey]
+            elif id_type == "show":
+                for m in main_id:
+                    if m in show_map:               show_map[m].append(item.ratingKey)
+                    else:                           show_map[m] = [item.ratingKey]
         util.print_end(length, f"Processed {len(items)} {'Movies' if library.is_movie else 'Shows'}")
         return movie_map, show_map
 
-    @retry(stop_max_attempt_number=6, wait_fixed=10000)
-    def get_guids(self, item):
-        return item.guids
-
-    def get_id(self, item, library, length):
-        expired = None
-        tmdb_id = None
-        imdb_id = None
-        tvdb_id = None
-        anidb_id = None
-        mal_id = None
-        error_message = None
-        if self.Cache:
-            if library.is_movie:                            tmdb_id, expired = self.Cache.get_tmdb_id("movie", plex_guid=item.guid)
-            else:                                           tvdb_id, expired = self.Cache.get_tvdb_id("show", plex_guid=item.guid)
-            if not tvdb_id and library.is_show:
-                tmdb_id, expired = self.Cache.get_tmdb_id("show", plex_guid=item.guid)
-                anidb_id, expired = self.Cache.get_anidb_id("show", plex_guid=item.guid)
-        if expired or (not tmdb_id and library.is_movie) or (not tvdb_id and not tmdb_id and library.is_show):
-            guid = requests.utils.urlparse(item.guid)
-            item_type = guid.scheme.split(".")[-1]
-            check_id = guid.netloc
-
-            if item_type == "plex":
-                tmdb_id = []
-                imdb_id = []
-                tvdb_id = []
-                if check_id == "movie":
-                    try:
-                        for guid_tag in self.get_guids(item):
-                            url_parsed = requests.utils.urlparse(guid_tag.id)
-                            if url_parsed.scheme == "tmdb":                 tmdb_id.append(int(url_parsed.netloc))
-                            elif url_parsed.scheme == "imdb":               imdb_id.append(url_parsed.netloc)
-                    except requests.exceptions.ConnectionError:
-                        util.print_stacktrace()
-                        logger.error(f"{'Cache | ! |' if self.Cache else 'Mapping Error:'} {item.guid:<46} | No External GUIDs found for {item.title}")
-                        return None, None
-                elif check_id == "show":
-                    try:
-                        for guid_tag in self.get_guids(item):
-                            url_parsed = requests.utils.urlparse(guid_tag.id)
-                            if url_parsed.scheme == "tvdb":                 tvdb_id.append(int(url_parsed.netloc))
-                            elif url_parsed.scheme == "imdb":               imdb_id.append(url_parsed.netloc)
-                            elif url_parsed.scheme == "tmdb":               tmdb_id.append(int(url_parsed.netloc))
-                    except requests.exceptions.ConnectionError:
-                        util.print_stacktrace()
-                        logger.error(f"{'Cache | ! |' if self.Cache else 'Mapping Error:'} {item.guid:<46} | No External GUIDs found for {item.title}")
-                        return None, None
-            elif item_type == "imdb":                       imdb_id = check_id
-            elif item_type == "thetvdb":                    tvdb_id = int(check_id)
-            elif item_type == "themoviedb":                 tmdb_id = int(check_id)
-            elif item_type == "hama":
-                if check_id.startswith("tvdb"):             tvdb_id = int(re.search("-(.*)", check_id).group(1))
-                elif check_id.startswith("anidb"):          anidb_id = re.search("-(.*)", check_id).group(1)
-                else:                                       error_message = f"Hama Agent ID: {check_id} not supported"
-            elif item_type == "myanimelist":                mal_id = check_id
-            elif item_type == "local":                      error_message = "No match in Plex"
-            else:                                           error_message = f"Agent {item_type} not supported"
-
-            if not error_message:
-                if mal_id and not anidb_id:
-                    try:                                            anidb_id = self.Arms.mal_to_anidb(mal_id)
-                    except Failed:                                  pass
-                if anidb_id and not tvdb_id:
-                    try:                                            tvdb_id = self.Arms.anidb_to_tvdb(anidb_id)
-                    except Failed:                                  pass
-                if anidb_id and not imdb_id:
-                    try:                                            imdb_id = self.Arms.anidb_to_imdb(anidb_id)
-                    except Failed:                                  pass
-                if not tmdb_id and imdb_id:
-                    if isinstance(imdb_id, list):
-                        tmdb_id = []
-                        new_imdb_id = []
-                        for imdb in imdb_id:
-                            try:
-                                tmdb_id.append(self.TMDb.convert_imdb_to_tmdb(imdb))
-                                new_imdb_id.append(imdb)
-                            except Failed:
-                                if self.Trakt:
-                                    try:
-                                        tmdb_id.append(self.Trakt.convert_imdb_to_tmdb(imdb))
-                                        new_imdb_id.append(imdb)
-                                    except Failed:
-                                        continue
-                                else:
-                                    continue
-                        imdb_id = new_imdb_id
-                    else:
-                        try:                                            tmdb_id = self.TMDb.convert_imdb_to_tmdb(imdb_id)
-                        except Failed:                                  pass
-                        if not tmdb_id and self.Trakt:
-                            try:                                            tmdb_id = self.Trakt.convert_imdb_to_tmdb(imdb_id)
-                            except Failed:                                  pass
-                if not tmdb_id and tvdb_id and library.is_show:
-                    try:                                            tmdb_id = self.TMDb.convert_tvdb_to_tmdb(tvdb_id)
-                    except Failed:                                  pass
-                    if not tmdb_id and self.Trakt:
-                        try:                                            tmdb_id = self.Trakt.convert_tvdb_to_tmdb(tvdb_id)
-                        except Failed:                                  pass
-                if not imdb_id and tmdb_id and library.is_movie:
-                    try:                                            imdb_id = self.TMDb.convert_tmdb_to_imdb(tmdb_id)
-                    except Failed:                                  pass
-                    if not imdb_id and self.Trakt:
-                        try:                                            imdb_id = self.Trakt.convert_tmdb_to_imdb(tmdb_id)
-                        except Failed:                                  pass
-                if not imdb_id and tvdb_id and library.is_show:
-                    try:                                            imdb_id = self.TMDb.convert_tvdb_to_imdb(tvdb_id)
-                    except Failed:                                  pass
-                    if not imdb_id and self.Trakt:
-                        try:                                            imdb_id = self.Trakt.convert_tvdb_to_imdb(tvdb_id)
-                        except Failed:                                  pass
-                if not tvdb_id and library.is_show:
-                    if tmdb_id:
-                        try:                                            tvdb_id = self.TMDb.convert_tmdb_to_tvdb(tmdb_id)
-                        except Failed:                                  pass
-                        if not tvdb_id and self.Trakt:
-                            try:                                            tvdb_id = self.Trakt.convert_tmdb_to_tvdb(tmdb_id)
-                            except Failed:                                  pass
-                    if not tvdb_id and imdb_id:
-                        try:                                            tvdb_id = self.TMDb.convert_imdb_to_tvdb(imdb_id)
-                        except Failed:                                  pass
-                        if not tvdb_id and self.Trakt:
-                            try:                                            tvdb_id = self.Trakt.convert_imdb_to_tvdb(tmdb_id)
-                            except Failed:                                  pass
-
-                if (not tmdb_id and library.is_movie) or (not tvdb_id and not (anidb_id and tmdb_id) and library.is_show):
-                    service_name = "TMDb ID" if library.is_movie else "TVDb ID"
-
-                    if self.Trakt:                                  api_name = "TMDb or Trakt"
-                    else:                                           api_name = "TMDb"
-
-                    if tmdb_id and imdb_id:                         id_name = f"TMDb ID: {tmdb_id} or IMDb ID: {imdb_id}"
-                    elif imdb_id and tvdb_id:                       id_name = f"IMDb ID: {imdb_id} or TVDb ID: {tvdb_id}"
-                    elif tmdb_id:                                   id_name = f"TMDb ID: {tmdb_id}"
-                    elif imdb_id:                                   id_name = f"IMDb ID: {imdb_id}"
-                    elif tvdb_id:                                   id_name = f"TVDb ID: {tvdb_id}"
-                    else:                                           id_name = None
-
-                    if anidb_id and not tmdb_id and not tvdb_id:    error_message = f"Unable to convert AniDB ID: {anidb_id} to TMDb ID or TVDb ID"
-                    elif id_name:                                   error_message = f"Unable to convert {id_name} to {service_name} using {api_name}"
-                    else:                                           error_message = f"No ID to convert to {service_name}"
-            if self.Cache and ((tmdb_id and library.is_movie) or ((tvdb_id or (anidb_id and tmdb_id)) and library.is_show)):
-                if not isinstance(tmdb_id, list):               tmdb_id = [tmdb_id]
-                if not isinstance(imdb_id, list):               imdb_id = [imdb_id]
-                if not isinstance(tvdb_id, list):               tvdb_id = [tvdb_id]
-                try:                                            tvdb_value = tvdb_id[0]
-                except IndexError:                              tvdb_value = None
-                for i in range(len(tmdb_id)):
-                    try:                                            imdb_value = imdb_id[i]
-                    except IndexError:                              imdb_value = None
-                    util.print_end(length, f"Cache | {'^' if expired is True else '+'} | {item.guid:<46} | {tmdb_id[i] if tmdb_id[i] else 'None':<6} | {imdb_value if imdb_value else 'None':<10} | {tvdb_value if tvdb_value else 'None':<6} | {anidb_id if anidb_id else 'None':<5} | {item.title}")
-                    self.Cache.update_guid("movie" if library.is_movie else "show", item.guid, tmdb_id[i], imdb_value, tvdb_value, anidb_id, expired)
-
-        if tmdb_id and library.is_movie:                return "movie", tmdb_id
-        elif tvdb_id and library.is_show:               return "show", tvdb_id
-        elif anidb_id and tmdb_id:                      return "movie", tmdb_id
-        else:
-            util.print_end(length, f"{'Cache | ! |' if self.Cache else 'Mapping Error:'} {item.guid:<46} | {error_message} for {item.title}")
-            return None, None
